add service layer

+++ Backend/types/job_types.go
@@ -3,8 +3,8 @@ package types
 import "time"
 
 type Job struct {
-	ID          uint64    `json:"id"`
-	UserID      uint      `json:"user_id"`
+	ID          int       `json:"id"`
+	UserID      int       `json:"user_id"`
 	Title       string    `json:"title"`
 	Company     string    `json:"company"`
 	Skills      []string  `json:"skills"`
@@ -16,7 +16,7 @@ type Job struct {
 }
 
 type JobReq struct {
-	UserID      uint     `json:"user_id"`
+	UserID      int      `json:"user_id"`
 	Title       string   `json:"title"`
 	Company     string   `json:"company"`
 	Skills      []string `json:"skills"`
